backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
ollama==0.1.5
pydantic==2.5.0
python-dotenv==1.0.0
cors==1.0.1
vercel-blob==0.0.5
requests==2.31.0
aiohttp==3.9.1
